# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Minha API",
        "description": "Documentação da API",
        "contact": {
            "email": "seu-email@dominio.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor principal"
        }
    ],
    "paths": {
...
        "/api/type_installation": {
            "get": {
                "tags": [
                    "public"
                ],
                "summary": "Obter lista de Tipos",
                "description": "Retorna uma lista de tipos de instalação",
                "operationId": "typeInstallationList",
                "responses": {
                    "200": {
                        "description": "Lista de exemplos retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TypeInstallation"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "public"
                ],
                "summary": "Inseri um tipo de intalação",
                "description": "Cria um novo recurso no banco de dados",
                "operationId": "typeInstallationStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Exemplo 1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Exemplo criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TypeInstallation"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/type_installation/{id}": {
            "put": {
                "tags": [
                    "public"
                ],
                "summary": "Atualizar o Tipo de Instalação",
                "description": "Atualiza um recurso existente com base no ID",
                "operationId": "typeInstallationUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do Tipo de Instalação",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Novo Nome do Exemplo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Exemplo atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TypeInstallation"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Exemplo não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "public"
                ],
                "summary": "Deletar um recurso",
                "description": "Deleta um recurso baseado no ID fornecido",
                "operationId": "typeInstallationDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do recurso que será deletado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recurso deletado com sucesso"
                    },
                    "404": {
                        "description": "Recurso não encontrado"
                    }
                }
            }
        },
        "/api/uf": {
            "get": {
                "tags": [
                    "public"
                ],
                "summary": "Obter lista de Ufs",
                "description": "Retorna uma lista",
                "operationId": "UfList",
                "responses": {
                    "200": {
                        "description": "Lista de exemplos retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Uf"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "public"
                ],
                "summary": "Inseri um objeto",
                "description": "Cria um novo recurso no banco de dados",
                "operationId": "UfStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "uf"
                                ],
                                "properties": {
                                    "uf": {
                                        "type": "string",
                                        "example": "SP"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Exemplo criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Uf"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/uf/{id}": {
            "put": {
                "tags": [
                    "public"
                ],
                "summary": "Atualizar um objeto",
                "description": "Atualiza um recurso existente com base no ID",
                "operationId": "UfUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do UF",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "uf"
                                ],
                                "properties": {
                                    "uf": {
                                        "type": "string",
                                        "example": "Novo Nome do Exemplo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Exemplo atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Uf"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Exemplo não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "public"
                ],
                "summary": "Deletar um recurso",
                "description": "Deleta um recurso baseado no ID fornecido",
                "operationId": "UfDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do recurso que será deletado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recurso deletado com sucesso"
                    },
                    "404": {
                        "description": "Recurso não encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "TypeInstallation": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Exemplo"
                    }
                },
                "type": "object"
            },
            "Uf": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "uf": {
                        "type": "string",
                        "example": "Sp"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}